feat(UI): port flatten to new UI
