+++ template_compare.go
