+++ types.go
@@ -198,12 +198,3 @@ func (app *App) Run() error {
 	}
 	return http.ListenAndServe(fmt.Sprintf("%s:%d", app.host, app.port), mux)
 }
-
-func (app *App) renderCompareForm(w http.ResponseWriter, r *http.Request) {
-	tmpl := template.Must(template.New("compare").Parse(compareFormTemplate))
-	err := tmpl.Execute(w, nil)
-	if err != nil {
-		app.logger.Error("failed to render upload page template", "err", err)
-		http.Error(w, "Failed to render page", http.StatusInternalServerError)
-	}
-}
